@@ -19,7 +19,9 @@
     "css-loader": "^0.9.1",
     "file-loader": "^0.8.1",
     "jsx-loader": "^0.12.2",
-    "react": "^0.13.1",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-bootstrap": "^0.31.0",
     "react-hot-loader": "^1.2.4",
     "style-loader": "^0.8.3",
     "url-loader": "^0.5.5",
