@@ -1,9 +1,8 @@
 import jQuery from "jquery";
+import ReactDOM from "react-dom";
 import React from "react";
 import ServerStatus from "jsx/server-status";
-import ClusterForm from "jsx/cluster-form";
-import ClusterList from "jsx/cluster-list";
-import Sidebar from "jsx/sidebar";
+import ClusterScreen from "jsx/cluster-screen";
 import {
   statusObservableTimer,
   addClusterSubject, addClusterResult, deleteClusterSubject, deleteClusterResult,
@@ -12,27 +11,24 @@ import {
 
 jQuery(document).ready(function($){
 
-  React.render(
-    React.createElement(ServerStatus, {statusObservableTimer}),
-    document.getElementById('cr-server-status')
-  );
+ $.urlParam = function(name){
+    var results = new RegExp('[\?&]' + name + '=([^&#]*)').exec(window.location.href);
+    if (results != null) {
+      return results[1] || 0;
+    } 
+    else {
+      return null;
+    }
+  }
 
-  React.render(
-    React.createElement(Sidebar, {clusterNames}),
-    document.getElementById('cr-sidebar')
-  );
+  let currentCluster: string = $.urlParam('currentCluster');
+  if(!currentCluster) {
+    currentCluster = 'all';
+  } 
 
-  React.render(
-    React.createElement(ClusterForm, {addClusterSubject, addClusterResult}),
-    document.getElementById('cr-cluster-form')
-  );
-
-  React.render(
-    React.createElement(ClusterList, {
-      clusterNames,
-      deleteSubject: deleteClusterSubject,
-      deleteResult: deleteClusterResult
-    }),
-    document.getElementById('cr-cluster-list')
+  ReactDOM.render(
+    React.createElement(ClusterScreen, {clusterNames, addClusterSubject, addClusterResult, currentCluster, deleteSubject: deleteClusterSubject,
+    deleteResult: deleteClusterResult, statusObservableTimer}),
+    document.getElementById('wrapper')
   );
 });
