@@ -1,6 +1,10 @@
 import React from "react";
 
 const sidebar = React.createClass({
+  propTypes: {
+    currentCluster: React.PropTypes.string.isRequired,
+    clusterNames: React.PropTypes.object    
+  },
 
   render: function() {
 
@@ -9,13 +13,13 @@ const sidebar = React.createClass({
           <div className="sidebar-nav navbar-collapse">
               <ul className="nav" id="side-menu">
                   <li>
-                      <a href="index.html"><i className="fa fa-sitemap fa-fw"></i> Cluster</a>
+                      <a href={'index.html?currentCluster=' + this.props.currentCluster}><i className="fa fa-sitemap fa-fw"></i> Cluster</a>
                   </li>
                   <li className="active">
-                      <a href="schedules.html"><i className="fa fa-calendar fa-fw"></i> Schedules</a>
+                      <a href={'schedules.html?currentCluster=' + this.props.currentCluster}><i className="fa fa-calendar fa-fw"></i> Schedules</a>
                   </li>
                   <li>
-                      <a href="repair.html"><i className="fa fa-wrench fa-fw"></i> Repair</a>
+                      <a href={'repair.html?currentCluster=' + this.props.currentCluster}><i className="fa fa-wrench fa-fw"></i> Repair</a>
                   </li>
 
               </ul>
